backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0

llama-index-core==0.10.68
llama-index-llms-groq==0.1.3

llama-index-embeddings-fastembed
fastembed
